// client/src/Redux/Action/types.ts
enum ActionType {
  DISPLAY_ALERT = "DISPLAY_ALERT",
  CLEAR_ALERT = "CLEAR_ALERT",

  SETUP_USER_BEGIN = "SETUP_USER_BEGIN",
  SETUP_USER_SUCCESS = "SETUP_USER_SUCCESS",
  SETUP_USER_ERROR = "SETUP_USER_ERROR",

  LOGOUT_USER = "LOGOUT_USER",

  UPDATE_USER_BEGIN = "UPDATE_USER_BEGIN",
  UPDATE_USER_SUCCESS = "UPDATE_USER_SUCCESS",
  UPDATE_USER_ERROR = "UPDATE_USER_ERROR",

  HANDLE_CHANGE = "HANDLE_CHANGE",
  CLEAR_VALUES = "CLEAR_VALUES",

  CREATE_STATION_BEGIN = "CREATE_STATION_BEGIN",
  CREATE_STATION_SUCCESS = "CREATE_STATION_SUCCESS",
  CREATE_STATION_ERROR = "CREATE_STATION_ERROR",

  GET_STATION_BEGIN = "GET_STATIONS_BEGIN",
  GET_STATION_SUCCESS = "GET_STATIONS_SUCCESS",

  SET_EDIT_STATION = "SET_EDIT_STATION",

  DELETE_STATION_BEGIN = "DELETE_STATION_BEGIN",
  DELETE_STATION_ERROR = "DELETE_STATION_ERROR",

  EDIT_STATION_BEGIN = "EDIT_STATION_BEGIN",
  EDIT_STATION_SUCCESS = "EDIT_STATION_SUCCESS",
  EDIT_STATION_ERROR = "EDIT_STATION_ERROR",

  SHOW_STATION_BEGIN = "SHOW_STATION_BEGIN",
  SHOW_STATION_SUCCESS = "SHOW_STATION_SUCCESS",
  CLEAR_FILTERS = "CLEAR_FILTERS",
  CHANGE_PAGE = "CHANGE_PAGE",

  GET_CURRENT_USER_BEGIN = "GET_CURRENT_USER_BEGIN",
  GET_CURRENT_USER_SUCCESS = "GET_CURRENT_USER_SUCCESS",
}

export default ActionType